pyproject.toml
[project]
name = "wakebuilder"
version = "0.1.0"
description = "A comprehensive training platform for creating custom wake word detection models locally"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Sami RAJICHI"}
]
license = {text = "Apache-2.0"}
keywords = ["wake-word", "speech-recognition", "voice-assistant", "machine-learning", "audio-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "cn2an>=0.5.23",
    "coqui-tts>=0.27.3",
    "datasets==2.0.0",
    "edge-tts>=7.2.7",
    "evaluate>=0.4.6",
    "fastapi>=0.121.3",
    "fugashi>=1.5.2",
    "hf-xet>=1.2.0",
    "jaconv>=0.4.1",
    "jieba>=0.42.1",
    "kokoro>=0.9.4",
    "librosa>=0.11.0",
    "mkdocs-material>=9.7.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mojimoji>=0.0.13",
    "noisereduce>=3.0.3",
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "ordered-set>=4.1.0",
    "piper-tts>=1.3.0",
    "pyopenjtalk>=0.4.1",
    "pypinyin>=0.55.0",
    "python-multipart>=0.0.20",
    "soundfile>=0.13.1",
    "tensorflow>=2.20.0",
    "tensorflow-hub>=0.16.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "unidic-lite>=1.0.8",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.6",
    "types-requests>=2.32.4.20260107",
]
